提交文件

+++ src/main/java/com/ixcoret/blog/controller/CategoryController.java
@@ -1,19 +1,19 @@
 package com.ixcoret.blog.controller;
 
-import com.ixcoret.blog.enums.ResultEnum;
-import com.ixcoret.blog.pojo.entity.Category;
-import com.ixcoret.blog.pojo.vo.CategoryBackVO;
-import com.ixcoret.blog.pojo.vo.form.CategoryForm;
+import com.ixcoret.blog.api.Page;
+import com.ixcoret.blog.api.Result;
+import com.ixcoret.blog.dto.CategoryDTO;
+import com.ixcoret.blog.dto.Condition;
+import com.ixcoret.blog.enums.ResultCodeEnum;
 import com.ixcoret.blog.service.CategoryService;
-import com.ixcoret.blog.utils.Result;
+import com.ixcoret.blog.vo.CategoryBackVO;
+import com.ixcoret.blog.vo.CategoryOptionVO;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.time.LocalDateTime;
 import java.util.List;
 
 /**
@@ -29,24 +29,24 @@ public class CategoryController {
 
     @PostMapping("/admin/categories")
     @ApiOperation("新增分类")
-    public Result save(@Valid @RequestBody CategoryForm categoryForm) {
-        Category category = new Category();
-        BeanUtils.copyProperties(categoryForm, category);
-        category.setCreateTime(LocalDateTime.now());
-        categoryService.save(category);
+    public Result save(@Valid @RequestBody CategoryDTO categoryDTO) {
+        categoryService.save(categoryDTO);
         return Result.success();
     }
 
     @PutMapping("/admin/categories")
     @ApiOperation("修改分类")
-    public Result update(@RequestBody CategoryForm categoryForm) {
-        categoryService.update(categoryForm);
+    public Result update(@Valid @RequestBody CategoryDTO categoryDTO) {
+        categoryService.update(categoryDTO);
         return Result.success();
     }
 
     @DeleteMapping("/admin/categories/{id}")
     @ApiOperation("根据id删除")
     public Result deleteById(@PathVariable Integer id) {
+        if (id == null || id < 1) {
+            return Result.error(ResultCodeEnum.PARAMS_ERROR);
+        }
         categoryService.deleteById(id);
         return Result.success();
     }
@@ -54,9 +54,8 @@ public class CategoryController {
     @DeleteMapping("/admin/categories")
     @ApiOperation("根据id批量删除")
     public Result deleteBatch(@RequestBody List<Integer> ids) {
-
-        if (ids == null || ids.size() == 0) {
-            return Result.error(ResultEnum.PARAMS_ERROR);
+        if (ids == null || ids.isEmpty()) {
+            return Result.error(ResultCodeEnum.PARAMS_ERROR);
         }
         categoryService.deleteBatch(ids);
         return Result.success();
@@ -64,8 +63,15 @@ public class CategoryController {
 
     @GetMapping("/admin/categories")
     @ApiOperation("获取后台分类列表")
-    public Result<List<CategoryBackVO>> listBackCategories() {
-        List<CategoryBackVO> backCategories = categoryService.listBackCategories();
-        return Result.success(backCategories);
+    public Result<Page<CategoryBackVO>> listBackCategories(Condition condition) {
+        Page<CategoryBackVO> page = categoryService.listBackCategories(condition);
+        return Result.success(page);
     }
+
+    @GetMapping("/admin/categories/options")
+    public Result<List<CategoryOptionVO>> listCategoryOptions() {
+        List<CategoryOptionVO> categoryOptions = categoryService.listCategoryOptions();
+        return Result.success(categoryOptions);
+    }
+
 }
